refactor(showcase): split each service to a module
I want to add several more tests, one for each service in `showcase`,
having all the tests in one module will get crowded pretty fast.
